+++ src/main/java/server/poptato/config/WebConfig.java
@@ -6,8 +6,9 @@ import org.springframework.http.HttpMethod;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import server.poptato.config.resolver.kakao.KakaoCodeResolver;
-import server.poptato.config.resolver.user.UserResolver;
+import server.poptato.external.kakao.resolver.KakaoCodeResolver;
+import server.poptato.external.kakao.resolver.OriginResolver;
+import server.poptato.user.resolver.UserResolver;
 
 import java.util.List;
 
@@ -16,6 +17,7 @@ import java.util.List;
 public class WebConfig implements WebMvcConfigurer {
     private final UserResolver userResolver;
     private final KakaoCodeResolver kakaoCodeResolver;
+    private final OriginResolver originResolver;
 
     @Override
     public void addCorsMappings(final CorsRegistry registry) {
@@ -31,6 +33,7 @@ public class WebConfig implements WebMvcConfigurer {
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
         resolvers.add(userResolver);
         resolvers.add(kakaoCodeResolver);
+        resolvers.add(originResolver);
 
     }
 }
