+++ src/main/java/server/poptato/external/kakao/resolver/KakaoCode.java
@@ -1,4 +1,4 @@
-package server.poptato.config.resolver.user;
+package server.poptato.external.kakao.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface UserId {
+public @interface KakaoCode {
 }
