+++ src/main/java/server/poptato/external/kakao/resolver/OriginHeader.java
@@ -1,0 +1,11 @@
+package server.poptato.external.kakao.resolver;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface OriginHeader {
+}
