+++ src/main/java/server/poptato/user/resolver/UserId.java
@@ -1,4 +1,4 @@
-package server.poptato.config.resolver.kakao;
+package server.poptato.user.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface KakaoCode {
+public @interface UserId {
 }
